sources,worker: refactor: extract shared fetch_options_with_ssh git2 helper
